Dev (#26)
* revert to faster-whisper a2t, also file structure better (#10)

* lots, work in prog

* nothing

* writer copy/paste whisper

* working currently

* restructure folder

* tweak to docker

* small small terraform/aws

* update requirements.txt file location

* docker bugz

* twaks to vastai-master

* docker setup.sh

* non-updates

* newline windows/unix update

* force added .gitkeep

* added cpu cors 4->8

* big frontend design rework/initial design

* updated idontfront, bigly

* images commit fix

* remove files

* updats to logging and the mini-image-compress

* added some logging

* update for .webp & docker file npm version issue

* img links update, tweaks to css idontfront

* update idontfront and small bug in preper

* REMOVED MINI-IMAGE.COM

* updated

* update idontfront

* nothing updated

* issue with docker fixed, failed git pull fixed

* logging update to transcriber

* nothing commit

* better logging vast.ai master, better transcriber SQL query

* updated

* config update

* updated to idontfront caching, minifier, and vars

* upated varz

* log

* created cloudfront function rewrite

* created cloudfront function rewrite

* try catch loop for updates preper

* update

* downsize prod

* updates, i dont remember

* sentiments in progress

* update sentiments, in prog

* update for less insane confusing channel and vods per channel numbers

* update

* update for downloader

* update vlaues of channels and vods per channel

* nothing really

* prod variables update

* logs maybe good, but 403 sub error

* logger updated @Transcriber

* transcribe. updating for big file a2t

* transcription update for infiity and too-big

* BIG, preper and downloader updated

* BIGLY updated microTranscriber, removed insanely-fast & other delete-mes

* small aws bug, 1 line

* 1 line bug

* 1 line bug

* selenium fixed, somehow broke

* github-action / workflows

* github-action / workflows

* github-action / workflows

* workflow

* workflow

* workflow

* workflow

* workflow

* small updates everywhere, preper is better

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* tiny non-update

* update

* renamed a docker-transciber file

* update

* update for emailer on mater

* ffmpeg pipe magic working locally, pray it works in lambda

* update

* updated emailer message, averages

* updates everywhere

* update email thingy

* update email thingy

* tweaks

* update conf

* updates to stuff idk

* emailer should be working

* update

* probably working

* idkman

* update

+++ SSScrapey-rework/src/controllers/MicroTranscriber/transcriber.py
@@ -20,12 +20,12 @@ import urllib.request
 import logging
 from utils.logging_config import LoggerConfig
 from controllers.MicroTranscriber.utils import TOO_BIG_LENGTH
-from utils.emailer import MetadataShitty
+
 
 # logger = Cloudwatch.log
 def logger():
     pass
-logger: logging.Logger = LoggerConfig("micro", env_varz.WHSP_IS_CLOUDWATCH == "True").get_logger()
+logger: logging.Logger = LoggerConfig("micro").get_logger()
 
 
 def getConnectionDb():
@@ -150,7 +150,8 @@ def setCompletedStatusDb(transcripts_s3_key_arr: List[str], vod: Vod):
     connection = getConnectionDb()
     t_status = "completed"
     transcripts_keys = json.dumps(transcripts_s3_key_arr)
-    logger.debug("   (setCompletedStatusDb) transcripts_keys:" + str( transcripts_keys))
+    # logger.debug("   (setCompletedStatusDb) transcripts_keys:" + str( transcripts_keys))
+    logger.debug(f"Completed: {vod.channels_name_id} - {vod.id} - {vod.title[:30]}")
     try:
         with connection.cursor() as cursor:
             sql = """
@@ -190,6 +191,9 @@ def downloadAudio(vod: Vod):
     audio_name = os.path.basename(audio_url)  # A trick to get the file name. eg) audio_url="https://[...].com/Calculated-v5057810.mp3" ---> audio_name="Calculated-v5057810.mp3"
     
     relative_filename = WHSP_A2T_ASSETS_AUDIO +  audio_name
+    relative_filename_aux = relative_filename.rsplit(".opus", 1)
+    relative_filename_mp4 = ".mp4".join(relative_filename_aux)
+    logger.debug("relative_filename_mp4------------- " +   relative_filename_mp4)
     logger.debug("audio_url: " + str( audio_url))
     logger.debug("audio_name: " + str( audio_name))
     
@@ -197,6 +201,10 @@ def downloadAudio(vod: Vod):
     if os.path.exists(relative_filename):
         logger.debug(f"File already exists: {relative_filename}")
         return relative_filename
+    
+    if os.path.exists(relative_filename_mp4):
+        logger.debug(f"File already exists: {relative_filename_mp4}")
+        return relative_filename_mp4
 
     try:
         relative_path, headers  = urllib.request.urlretrieve(audio_url, relative_filename) # audio_url = Calculated-v123123.ogg
@@ -214,9 +222,9 @@ def downloadAudio(vod: Vod):
 
 def uploadCaptionsToS3(saved_caption_files: List[str], vod: Vod):
     logger.debug("XXXXXXXXXXXXXX  uploadCaptionsToS3  XXXXXXXXXXXXXX")
-    logger.debug("    (uploadCaptionsToS3) channel: " + vod.channels_name_id)
-    logger.debug("    (uploadCaptionsToS3) vod.id: " + vod.id) 
-    logger.debug("    (uploadCaptionsToS3) vod.title: " + vod.title) 
+    logger.debug("    channel: " + vod.channels_name_id)
+    logger.debug("    vod.id: " + vod.id) 
+    logger.debug("    vod.title: " + vod.title) 
 
     s3 = boto3.client('s3')
     transcripts_s3_key_arr = []
@@ -262,52 +270,3 @@ def deleteAudioLocally(relative_path: str):
         logger.error(str(e))
     return 
 
-
-def getFromFancyMap(d: dict[int, list]):
-    # https://chatgpt.com/c/69278b8e-856c-8332-9475-66ce608d2298
-    # data: Dict[int, List[Vod]] = {
-    #     1: [a1,a2,a3],
-    #     2: [b1,b2,b3,b4,b5,b6],
-    #     3: [c1,c2],
-    # }
-    # outputs -> 1 column, 2nd column, 3rd, ....
-    # outputs -> a1,b1,c1, a2,b2,c2, a3,b3, b4, b5, b6   
-    keys = d.keys()
-    max_len = max(len(d[k]) for k in keys)
-    for x in range(max_len): # x = column
-        for y in keys: # y = row
-            row = d[y]
-            if x < len(row):
-                yield row[x] 
-
-def convertToFancyMap(vod_list: List[Vod]) -> Dict[int, List[Vod]]:
-    sub_list = []
-    magical_ordered_map = {}
-    previous = vod_list[0].channels_name_id # initialize
-    idx_rank = 0
-    for i, vod in enumerate(vod_list):
-        current = vod.channels_name_id
-        if previous != current:
-            magical_ordered_map[idx_rank] = sub_list
-            previous = current
-            sub_list = []
-            idx_rank += 1
-        sub_list.append(vod)
-    
-    # Add the last group
-    magical_ordered_map[idx_rank + 1] = sub_list
-
-    mega_count = 0
-    for key, v_list in magical_ordered_map.items():
-        logger.debug(f"----------- {str(key)} -----------")
-        for i, v in enumerate(v_list):
-            v: Vod = v
-            # logger.debug(i, " - ", v.channels_name_id, v.id, v.stream_date)
-            logger.debug(f"{i} - {v.channels_name_id} {v.id} {v.stream_date}")
-            mega_count += 1
-    logger.debug("Total = " + str(mega_count))
-    return magical_ordered_map
-
-    # "Transformers now supports natively BetterTransformer optimizations ... no need to use `model.to_bettertransformers()` Details: https://huggingface.co/docs/transformers/perf_infer_gpu_one#flashattention-and-memory-efficient-attention-through-pytorchs-scaleddotproductattention."
-    # if not is_flash_attn_2_available() or better_transformer:
-    #     pipe.model = pipe.model.to_bettertransformer()
